+++ CyRK/optimize/optimize_test.pyx
@@ -1,8 +1,10 @@
 # distutils: language = c++
 # cython: boundscheck=False, wraparound=False, nonecheck=False, cdivision=True, initializedcheck=False
 
+from libcpp cimport nullptr
 from libcpp.vector cimport vector
 
+from CyRK.cy.cysolver_api cimport CySolverDense
 from CyRK.optimize.brentq cimport c_brentq, OptimizeInfo
 
 # extra parameters
@@ -10,8 +12,10 @@ ctypedef struct extra_params:
     double[4] a
 
 # callback function
-cdef double f_example(double x, char *args_char_ptr) noexcept nogil:
+cdef double f_example(double x, double* y_ptr, char *args_char_ptr) noexcept nogil:
     cdef extra_params *args_ptr = <extra_params*> args_char_ptr
+    # y_ptr is unused in this example.
+
     # use Horner's method
     return ((args_ptr.a[3]*x + args_ptr.a[2])*x + args_ptr.a[1])*x + args_ptr.a[0]
 
@@ -35,4 +39,8 @@ def brentq_test(
         else:
             args_ptr.a[i] = args[i-1]
 
-    return c_brentq(f_example, xa, xb, xtol, rtol, mitr, args_vec, &optimize_info)
+    # Dense function is not used for this test; set to null
+    cdef CySolverDense* dense_func = NULL
+
+    # Call and return brentq root finder
+    return c_brentq(f_example, xa, xb, xtol, rtol, mitr, args_vec, &optimize_info, dense_func)